Refactors domain model tests
Restructures domain model tests by creating individual test files for each domain model.
This improves code organization and maintainability.
